safari-600.7-branch/LayoutTests/http/tests/security/referrer-policy-invalid.html

<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFramesAsText();
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<p>
This test checks an invalid referrer policy when navigating from an insecure
URL to another insecure URL. The test passes if the printed referrer is empty.
</p>
<iframe src="http://127.0.0.1:8000/security/resources/referrer-policy-start.html?invalid"></iframe>
</body>
</html>
