safari-600.7-branch/LayoutTests/http/tests/security/referrer-policy-invalid-expected.txt

CONSOLE MESSAGE: Failed to set referrer policy: The value 'invalid' is not one of 'always', 'default', 'never', or 'origin'. Defaulting to 'never'.
This test checks an invalid referrer policy when navigating from an insecure URL to another insecure URL. The test passes if the printed referrer is empty.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
HTTP Referer header is empty
Referrer is empty
